Move upload to utility process
Attempt to further lighten the load on the main thread
